requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
ale-py==0.11.2
array_record==0.8.3
astunparse==1.6.3
attrs==25.4.0
cachetools==6.2.2
certifi @ file:///home/task_176313480683104/conda-bld/certifi_1763135330277/work/certifi
chardet==5.2.0
charset-normalizer==3.4.4
chex==0.1.91
cloudpickle==3.1.2
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
decorator==5.2.1
dm-haiku==0.0.15
dm-tree==0.1.9
docstring_parser==0.17.0
dopamine==0.7
dopamine_rl==4.1.2
einops==0.8.1
etils==1.13.0
Farama-Notifications==0.0.4
filelock==3.20.0
flatbuffers==25.9.23
flax==0.12.1
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
gast==0.6.0
gin-config==0.5.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gym==0.25.2
gym-notices==0.1.0
gymnasium==1.2.2
h5py==3.15.1
humanize==4.14.0
idna==3.11
immutabledict==4.2.2
importlib_resources==6.5.2
jax==0.8.1
jax-cuda12-pjrt==0.8.1
jax-cuda12-plugin==0.8.1
jaxlib==0.8.1
Jinja2==3.1.6
jmp==0.0.4
keras==3.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optax==0.2.6
optree==0.18.0
orbax-checkpoint==0.11.28
packaging==25.0
pandas==2.3.3
pillow==12.0.0
promise==2.3
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygame==2.6.1
pyglet==2.1.11
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scipy==1.16.3
setuptools==80.9.0
simple-parsing==0.1.7
simplejson==3.20.2
six==1.17.0
stable_baselines3==2.7.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-estimator==2.15.0
tensorflow-metadata==1.17.2
tensorflow-probability==0.25.0
tensorstore==0.1.79
termcolor==3.2.0
tf-slim==1.1.0
tf_keras==2.20.1
toml==0.10.2
toolz==1.1.0
torch==2.9.1
tqdm==4.67.1
treescope==0.1.10
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
xvfbwrapper==0.2.15
zipp==3.23.0
